Changed saveOnPermission() to take an extra parameter (prompt string ID) to make it more flexible.  Changed all calls to saveOnPermission().
Changed logout to call saveOnPermission() so that the user can choose to
save the current state when logging out.

Fixed logout to clear dirty flag.

Changed getCurrentAlbum() to select the first album if none is selected.  This
is to cut down on the possiblity that the method will return null.

	public void jListKeyPressed(KeyEvent e) {
    public void deleteSelectedPictures() {
    public void movePicturesUp() {
    public void movePicturesDown() {

Added
  deleteSelectedPictures(),
  movePicturesUp(), and
  movePicturesDown().
These call the coreUtil methods of the same name, but they also mark the
current document as dirty.  I changed the PictureInspector to use these
instead of the CoreUtils directly so that we can reliably track the dirty state
of the document.